refactor: Users Request/Response 클래스 위치 변경
