--- src/main/java/com/team2/finalproject/domain/users/model/dto/LoginResponse.java
@@ -1,16 +1,0 @@
-package com.team2.finalproject.domain.users.model.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class LoginResponse {
-
-    private String name;
-    private Long centerId;
-}
