+++ src/main/java/com/team2/finalproject/domain/users/model/dto/request/RegisterDriverRequest.java
@@ -1,0 +1,20 @@
+package com.team2.finalproject.domain.users.model.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class RegisterDriverRequest {
+
+    private long centerId;
+    private long smId;
+    private String name;
+    private String username;
+    private String password;
+    private String phoneNumber;
+}
